Fixed issues with futures not returning any results. Fixed order of results being combined.
Futures were not returning results because the outer loop in the render
function was decrementing high to low and the loop was initialized
low to high meaning the loop would just get skipped.
The order of the results was fixed because due to the results being
generated from high to low (image_height to 0) they are stored in
reverse. This is fixed by iterating in reverse and combining them.
As of this change, multithreading is working.